Framework In Place
Now the framework is in place for the actual app.  Now I need to get it
all up and running so that each view is passing data.  Once I have the
views passing data correctly from Parse, It will be time to make sure
Parse has all the data and is continually up to date.